+++ openbr/plugins/independent.cpp
@@ -129,6 +129,8 @@ class IndependentTransform : public MetaTransform
         return independentTransform;
     }
 
+    bool timeVarying() const { return transform->timeVarying(); }
+
     static void _train(Transform *transform, const TemplateList *data)
     {
         transform->train(*data);
@@ -168,6 +170,27 @@ class IndependentTransform : public MetaTransform
             dst.clear();
         }
         dst.append(mats);
+    }
+
+    void projectUpdate(const Template &src, Template &dst)
+    {
+        dst.file = src.file;
+        QList<Mat> mats;
+        for (int i=0; i<src.size(); i++) {
+            transforms[i%transforms.size()]->projectUpdate(Template(src.file, src[i]), dst);
+            mats.append(dst);
+            dst.clear();
+        }
+        dst.append(mats);
+    }
+
+    void projectUpdate(const TemplateList &src, TemplateList &dst)
+    {
+        dst.reserve(src.size());
+        foreach (const Template &t, src) {
+            dst.append(Template());
+            projectUpdate(t, dst.last());
+        }
     }
 
     void store(QDataStream &stream) const