+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.4.0 - ??/??/??
 ================
+* Added -evalDetection and -plotDetection for evaluating and plotting object detection accuracy (#9)
 
 0.3.0 - 5/22/13
 ===============